Replaced outdated VAProfile/Models with the new models in VAProfile/Models/V3
