+++ app/controllers/v0/profile/address_validation_controller.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'va_profile/models/v3/validation_address'
+require 'va_profile/models/validation_address'
 require 'va_profile/address_validation/v3/service'
 
 module V0
@@ -11,7 +11,7 @@ module V0
       skip_before_action :authenticate, only: [:create]
 
       def create
-        address = VAProfile::Models::V3::ValidationAddress.new(address_params)
+        address = VAProfile::Models::ValidationAddress.new(address_params)
         raise Common::Exceptions::ValidationErrors, address unless address.valid?
 
         Rails.logger.warn('AddressValidationController#create request completed', sso_logging_info)
