+++ app/swagger/swagger/schemas/vet360/contact_information.rb
@@ -2,7 +2,7 @@
 
 require 'va_profile/contact_information/v2/person_response'
 require 'va_profile/contact_information/v2/service'
-require 'va_profile/models/v3/address'
+require 'va_profile/models/address'
 require 'va_profile/models/telephone'
 require 'common/models/redis_store'
 require 'common/models/concerns/cache_aside'
