+++ lib/va_profile/models/person.rb
@@ -4,7 +4,6 @@ require_relative 'address'
 require_relative 'base'
 require_relative 'email'
 require_relative 'telephone'
-require_relative 'permission'
 
 module VAProfile
   module Models
@@ -14,10 +13,10 @@ module VAProfile
       attribute :emails, Email, array: true
       attribute :source_date, Vets::Type::ISO8601Time
       attribute :telephones, Telephone, array: true
-      attribute :permissions, Permission, array: true
       attribute :transaction_id, String
       attribute :updated_at, Vets::Type::ISO8601Time
       attribute :vet360_id, String
+      attribute :va_profile_id, String
 
       # Converts a decoded JSON response from VAProfile to an instance of the Person model
       # @param body [Hash] the decoded response body from VAProfile
@@ -27,7 +26,6 @@ module VAProfile
         addresses = body['addresses']&.map { |a| VAProfile::Models::Address.build_from(a) }
         emails = body['emails']&.map { |e| VAProfile::Models::Email.build_from(e) }
         telephones = body['telephones']&.map { |t| VAProfile::Models::Telephone.build_from(t) }
-        permissions = body['permissions']&.map { |t| VAProfile::Models::Permission.build_from(t) }
 
         VAProfile::Models::Person.new(
           created_at: body['create_date'],
@@ -37,8 +35,8 @@ module VAProfile
           addresses: addresses || [],
           emails: emails || [],
           telephones: telephones || [],
-          permissions: permissions || [],
-          vet360_id: body['vet360_id']
+          vet360_id: body['vet360_id'] || body['va_profile_id'],
+          va_profile_id: body['vet360_id'] || body['va_profile_id']
         )
       end
     end
