+++ lib/va_profile/models/validation_address.rb
@@ -6,29 +6,27 @@ require 'common/hash_helpers'
 module VAProfile
   module Models
     # Model for addresses sent and received from the VA profile address validation API
+    # AddressValidationV is used for ProfileServiceV3 and ContactInformationV2
     class ValidationAddress < BaseAddress
       # Convert a ValidationAddress into a hash that can be sent to the address validation
       # API
       # @return [Hash] hash that is formatted for POSTing to address validation API
       def address_validation_req
         Common::HashHelpers.deep_remove_blanks(
-          requestAddress: attributes.slice(
+          address: attributes.slice(
             :address_line1,
             :address_line2,
-            :address_line3,
-            :city,
-            :international_postal_code
+            :address_line3
           ).deep_transform_keys { |k| k.to_s.camelize(:lower) }.merge(
-            requestCountry: {
-              countryCode: @country_code_iso3
-            },
-            addressPOU: @address_pou,
-            stateProvince: {
-              code: @state_code,
-              name: @province
-            },
+            intPostalCode: @international_postal_code,
+            cityName: @city,
             zipCode5: @zip_code,
-            zipCode4: @zip_code_suffix
+            zipCode4: @zip_code_suffix,
+            country: { countryCodeISO3: @country_code_iso3,
+                       countryName: @country_name, countryCodeFIPS: @country_code_fips },
+            state: { stateCode: @state_code, stateName: @state_name },
+            province: { provinceName: @province_name, provinceCode: @province_code },
+            addressPOU: @address_pou
           )
         )
       end
@@ -36,17 +34,16 @@ module VAProfile
       # @return [VAProfile::Models::ValidationAddress] validation address model created from
       #   address validation API response
       def self.build_from_address_suggestion(address_suggestion_hash)
-        address_hash = address_suggestion_hash['address']
-        address_type = address_suggestion_hash['address_meta_data']['address_type'].upcase
+        address_type = address_suggestion_hash['address_type'].upcase
         attributes = {
-          address_line1: address_hash['address_line1'],
-          address_line2: address_hash['address_line2'],
-          address_line3: address_hash['address_line3'],
+          address_line1: address_suggestion_hash['address_line1'],
+          address_line2: address_suggestion_hash['address_line2'],
+          address_line3: address_suggestion_hash['address_line3'],
           address_type:,
-          city: address_hash['city'],
-          country_name: address_hash['country']['name'],
-          country_code_iso3: address_hash['country']['iso3_code']
-        }.merge(regional_attributes(address_type, address_hash))
+          city: address_suggestion_hash['city_name'],
+          country_name: address_suggestion_hash.dig('country', 'country_name'),
+          country_code_iso3: address_suggestion_hash.dig('country', 'country_code_iso3')
+        }.merge(regional_attributes(address_type, address_suggestion_hash))
 
         new(attributes)
       end
@@ -54,17 +51,27 @@ module VAProfile
       def self.regional_attributes(address_type, address_hash)
         if address_type == INTERNATIONAL
           {
-            province: address_hash['state_province']['name'],
-            international_postal_code: address_hash['international_postal_code']
+            province: address_hash['province']['province_name'],
+            international_postal_code: address_hash['int_postal_code']
           }
         else
           {
-            state_code: address_hash['state_province']['code'],
-            county_code: address_hash.dig('county', 'county_fips_code'),
-            county_name: address_hash.dig('county', 'name'),
-            zip_code: address_hash['zip_code5'], zip_code_suffix: address_hash['zip_code4']
+            state_code: address_hash.dig('state', 'state_code'),
+            state_name: address_hash.dig('state', 'state_name'),
+            county_code: address_hash.dig('county', 'county_code'),
+            county_name: address_hash.dig('county', 'county_name'),
+            zip_code: address_hash['zip_code5'],
+            zip_code_suffix: address_hash['zip_code4']
           }
         end
+      end
+
+      def self.build_address_metadata(address_suggestion_hash)
+        {
+          confidence_score: address_suggestion_hash['confidence'],
+          address_type: address_suggestion_hash['address_type'],
+          delivery_point_validation: address_suggestion_hash['delivery_point_validation']
+        }
       end
     end
   end
