+++ modules/ask_va_api/spec/requests/ask_va_api/v0/address_validation_spec.rb
@@ -4,9 +4,9 @@ require 'rails_helper'
 
 RSpec.describe 'AskVAApi::V0::AddressValidation', type: :request do
   let(:user) { build(:user) }
-  let(:multiple_match_addr) { build(:va_profile_v3_address, :multiple_matches) }
-  let(:invalid_address) { build(:va_profile_v3_validation_address).to_h }
-  let(:incorrect_address_pou) { build(:va_profile_v3_address, :incorrect_address_pou) }
+  let(:multiple_match_addr) { build(:va_profile_address, :multiple_matches) }
+  let(:invalid_address) { build(:va_profile_validation_address).to_h }
+  let(:incorrect_address_pou) { build(:va_profile_address, :incorrect_address_pou) }
 
   shared_examples 'invalid address' do
     it 'returns an error' do
