+++ modules/mobile/spec/services/sync_update_service_spec.rb
@@ -16,7 +16,7 @@ describe Mobile::V0::Profile::SyncUpdateService do
 
   describe '#v2_save_and_await_response' do
     let(:user) { create(:user, :api_auth_v2, :legacy_icn) }
-    let(:params) { build(:va_profile_v3_address, :override, id: 577_127) }
+    let(:params) { build(:va_profile_address, :override, id: 577_127) }
 
     context 'when it succeeds' do
       let(:transaction) do
