+++ modules/simple_forms_api/app/models/form_engine/address.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module FormEngine
-  class Address < VAProfile::Models::V3::Address
+  class Address < VAProfile::Models::Address
     def initialize(params)
       super(params)
