+++ modules/veteran/lib/tasks/update_vso_addresses.rake
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'va_profile/models/v3/validation_address'
+require 'va_profile/models/validation_address'
 require 'va_profile/address_validation/v3/service'
 
 ADDRESS_BATCH1 = [
@@ -902,7 +902,7 @@ ADDRESS_BATCH5 = [
 # @param org [Hash] A hash containing the details of the organization's address.
 # @return [VAProfile::Models::ValidationAddress] A validation address object ready for address validation service.
 def build_validation_address(org)
-  validation_model = VAProfile::Models::V3::ValidationAddress
+  validation_model = VAProfile::Models::ValidationAddress
   validation_model.new(
     address_pou: org[:address_pou],
     address_line1: org[:address_line1],
@@ -920,7 +920,7 @@ end
 # @param candidate_address [VAProfile::Models::ValidationAddress] The address to be validated.
 # @return [Hash] The response from the address validation service.
 def validate_address(candidate_address)
-  validation_service = VAProfile::Models::V3::ValidationAddress
+  validation_service = VAProfile::Models::ValidationAddress
   validation_service.candidate(candidate_address)
 end
