+++ spec/factories/va_profile/persons.rb
@@ -3,10 +3,8 @@
 FactoryBot.define do
   factory :person, class: 'VAProfile::Models::Person' do
     addresses   { [build(:va_profile_address), build(:va_profile_address, :mailing)] }
-    emails      { [build(:email)] }
-    # TODO: test that these telephone & permission factories register
-    telephones  { [build(:telephone)] }
-    permissions { [build(:permission)] }
+    emails      { [build(:email, :contact_info_v2)] }
+    telephones  { [build(:telephone, :contact_info_v2)] }
     source_date { '2018-04-09T11:52:03-06:00' }
     created_at  { '2017-04-09T11:52:03-06:00' }
     updated_at  { '2017-04-09T11:52:03-06:00' }
