+++ spec/lib/common/models/concerns/cache_aside_spec.rb
@@ -6,10 +6,10 @@ require 'common/models/concerns/cache_aside'
 describe Common::CacheAside do
   let(:user) { build(:user, :loa3) }
 
-  let(:person) { build(:person_v2) }
+  let(:person) { build(:person) }
 
   before do
-    allow(VAProfile::Models::V3::Person).to receive(:build_from).and_return(person)
+    allow(VAProfile::Models::Person).to receive(:build_from).and_return(person)
   end
 
   describe '#do_cached_with' do
