+++ spec/lib/va_profile/models/address_spec.rb
@@ -11,7 +11,7 @@ describe VAProfile::Models::Address do
       expect(address.latitude).to eq(38.901)
       expect(address.longitude).to eq(-77.0347)
       expect(address.geocode_precision).to eq(100.0)
-      expect(address.geocode_date).to eq(Time.zone.parse('2018-04-13 17:01:18 UTC'))
+      expect(address.geocode_date).to eq(Time.zone.parse('2024-08-27T18:51:06.012Z'))
     end
   end
