+++ spec/lib/va_profile/models/validation_address_spec.rb
@@ -9,10 +9,19 @@ describe VAProfile::Models::ValidationAddress do
   describe '#address_validation_req' do
     it 'formats the address for an address validation request' do
       expect(address.address_validation_req).to match(
-        requestAddress: { 'addressLine1' => '37 1st st',
-                          'city' => 'Brooklyn', :requestCountry => { countryCode: 'USA' },
-                          :addressPOU => 'RESIDENCE/CHOICE', :stateProvince => { code: 'NY' },
-                          :zipCode5 => '11249' }
+        address: {
+          'addressLine1' => '37 1st st',
+          :cityName => 'Brooklyn',
+          :zipCode5 => '11249',
+          :country => {
+            countryCodeISO3: 'USA'
+          },
+          :state => {
+            stateCode: 'NY'
+          },
+          :province => {},
+          :addressPOU => 'RESIDENCE'
+        }
       )
     end
   end
@@ -25,31 +34,27 @@ describe VAProfile::Models::ValidationAddress do
     context 'with a domestic address' do
       let(:address_suggestion_hash) do
         {
-          'address' => {
-            'county' => { 'name' => 'Kings', 'county_fips_code' => '36047' },
-            'state_province' => { 'name' => 'New York', 'code' => 'NY' },
-            'country' => {
-              'name' => 'United States',
-              'code' => 'USA',
-              'fips_code' => 'US',
-              'iso2_code' => 'US',
-              'iso3_code' => 'USA'
-            },
-            'address_line1' => '37 N 1st St',
-            'city' => 'Brooklyn',
-            'zip_code5' => '11249',
-            'zip_code4' => '3939'
+          'address_line1' => '37 N 1st St',
+          'city_name' => 'Brooklyn',
+          'zip_code5' => '11249',
+          'zip_code4' => '3939',
+          'addressPOU' => 'RESIDENCE',
+          'county' => { 'county_name' => 'Kings', 'county_code' => '36047' },
+          'state' => { 'state_name' => 'New York', 'state_code' => 'NY' },
+          'country' => {
+            'country_name' => 'United States',
+            'country_code_fips' => 'US',
+            'country_code_iso3' => 'USA'
           },
           'geocode' => {
             'calc_date' => '2020-01-23T03:15:47+00:00',
-            'location_precision' => 31.0, 'latitude' => 40.717029,
+            'location_precision' => 31.0,
+            'latitude' => 40.717029,
             'longitude' => -73.964956
           },
-          'address_meta_data' => {
-            'confidence_score' => 100.0, 'address_type' => 'Domestic',
-            'delivery_point_validation' => 'UNDELIVERABLE',
-            'validation_key' => -646_932_106
-          }
+          'confidence' => 100.0,
+          'address_type' => 'Domestic',
+          'delivery_point_validation' => 'UNDELIVERABLE'
         }
       end
 
@@ -72,26 +77,26 @@ describe VAProfile::Models::ValidationAddress do
     context 'with an international address' do
       let(:address_suggestion_hash) do
         {
-          'address' => {
-            'address_line1' => '898 Broadway W',
-            'city' => 'Vancouver',
-            'international_postal_code' => 'V5Z 1J8',
-            'county' => {},
-            'state_province' => { 'name' => 'British Columbia', 'code' => 'BC' },
-            'country' => {
-              'name' => 'Canada', 'code' => 'CAN',
-              'fips_code' => 'CA', 'iso2_code' => 'CA', 'iso3_code' => 'CAN'
-            }
+          'address_line1' => '898 Broadway W',
+          'city_name' => 'Vancouver',
+          'int_postal_code' => 'V5Z 1J8',
+          'county' => {},
+          'province' => {
+            'province_name' => 'British Columbia',
+            'province_code' => 'BC'
+          },
+          'country' => {
+            'country_name' => 'Canada',
+            'country_code_fips' => 'CA',
+            'country_code_iso3' => 'CAN'
           },
           'geocode' => {
             'calc_date' => '2020-04-10T17:29:41Z',
             'location_precision' => 10.0, 'latitude' => 49.2635,
             'longitude' => -123.13873
           },
-          'address_meta_data' => {
-            'confidence_score' => 97.76, 'address_type' => 'International',
-            'validation_key' => -1_941_145_206
-          }
+          'confidence' => 97.76,
+          'address_type' => 'International'
         }
       end
