+++ spec/models/async_transaction/vet360/base_spec.rb
@@ -83,7 +83,7 @@ RSpec.describe AsyncTransaction::Vet360::Base, type: :model do
 
   describe '.start' do
     let(:user) { build(:user, :loa3, :legacy_icn) }
-    let(:address) { build(:va_profile_v3_address, source_system_user: user.icn) }
+    let(:address) { build(:va_profile_address, source_system_user: user.icn) }
 
     it 'returns an instance with the user uuid', :aggregate_failures do
       VCR.use_cassette('va_profile/v2/contact_information/post_address_success', VCR::MATCH_EVERYTHING) do
